[IMP] website_slides: Do not refresh after adding slide/section
This commit allows the user to add slides and sections in the front-end
without having to refresh the page. The slide, the section or both at a time
will directly be added to the dom after the rpc call.

This change allows us to completely remove the resequencing logic
in the add_slide controller and the slide.channel model.

Note that a few particular type of slides will still need a redirection
to the backend (webpage and quiz).

Also, some _rpc calls had to be replaced by ajax.jsonRpc due to the
_rpc calls not working at all for the newly added slide/section controls.

LINKS:

TaskID: 2089854
PR: #39242