[IMP] hw_drivers: Make iot_devices/bin/* exec
We want to add the possibility to add dependencies without having to
rebuild the image of the box. In `iot_devices` we can now  add a `bin`
directory. All of the files located in this directory will copied to
the Box and made executable so that they can be called from inside
drivers.

TaskID: 2082282