[IMP] website_sale: thumbnail of video as default image
After this commit is applied, this will add default product image according to
URL and type of access(private or public).There are some scenarios for type
of URL and program behaviour  in response to URL which are mentioned below.

If the provider doesn't provide a thumbnail (if video URL is not public)
in that case,  background image will set as default  and users can also add an
image manually. A miniature of the video is calculated based on the provider
and the available provider is Youtube, Vimeo, Dailymotion  and Instagram If one
want to add any other video provider than only regular expression and seprate
function  will be added.we provide miniature/thumbnail as a default  image
if video URL is not private otherwise, the background image will set as
the default image .

task-2154812
Closes #44537

Co-authored-by: jpr-odoo <jpr@openerp.com>