[IMP] mail, hr_holidays: add current user by default on new channel
Before this commit, when a user creates a private channel without adding himself
as a member of it, it raises an access error.

Also adapt "out of office" test to avoid creating the channel as OdooBot, and
ensure `out_of_office_date_end` is always added to the result for consistency.

task-2172420