[IMP] hr_holidays_attendance : add extra_hours deduction
It is now possible do deduct extra_hours from the total of it (which is
computed on the employee).
For each approved leave where deduct_from_extra_hours = True, its duration
is deducted from the total of extra_hours for the employee.

Being holidays officer now implies to be attendance officer.

TASK#1904850