[REF] hr_expense: convert "onchanges" to stored-editable computed fields
"copy=True" has to be added for some required computed fields.
"_onchange_product_id" (in HrExpense model) has been emptied but not removed as it is called by an inherited class in sale_expense module.
"_onchange_product_uom_id" (in HrExpense model) is used to raise an UserError and has been left as it is.