[REF] hr_holidays: add copy=True and readonly=False to computed fields converted from onchange
3 tests have been removed because they were trying to create a holiday with a "validate" state to test if a ValidationError was correctly raised.
As the field for the state is now computed with the required holiday_status_id field, it is not possible anymore to directly create a holiday with a given state.

That's also why 1 test has been modified to set the state to the desired value for purpose of the test.
The state has to be updated after the creation.