[IMP] base: currency rate depend on company
This was already the kind of the case because the company_id had a
default value, but it would be nice to ensure that it is always the
case. This allows more robust uses, and is clearer for the user as the
company's currency will always be at 1.0. This was not possible for
multi-company with different main currencies.