[FIX] project: Make 'Sub-tasks' feature editable on projects
Before this commit, it was not possible to enable/disable sub-tasks
feature for particular project.

Purpose of the commit https://github.com/odoo/odoo/commit/2ac06038dd08be863c27cc7d1d7aec9a70ec2105
was to enable subtasks per project but currently it's not due to
non stored compute field.

So now it is possible to enable/disable 'sub-tasks' per project and
if subtask setting is enabled and project is not fsm then default
value will be the True for new project.
Also on disable of feature from settings it will disable 'sub-task' for
all the project.

Taskid: 2192763
Related: https://github.com/odoo/enterprise/pull/8468
Closes: #45887