[IMP] website_slides(_survey): allow to keep user progress after leaving course
PURPOSE

Before this commit when a user left a course, his progress was lost and he had
to retake the whole course from the start. It also allowed people from gaining
karma easily by leaving the course and taking the quiz again and again.
After this commit the user will keep his progress event after leaving the
course and rejoining it later.

SPECIFICATION

When leaving the course the corresponding 'slide.channel.partner' record is
deleted but the 'slide.slide.partner' records are kept unchanged. So when
joining the course a new 'slide.channel.partner' is create again.
On the other hand, after failing a survey all the records are deleted so the
user has to retake the whole certification course.

LINKS

Task ID : 2199207
PR : #46181