[FIX] mail: Allow to disable the attachment box in chatter
Currently there is option in chatter as 'disable_attachment_box' which
is going to disable attachment in chatter but it was not passed from
the form view renderer so functionality does not work for the form view.

After this commit user will able to disable the attachment in chatter of
the form view by passing the options on 'oe_chatter' div.
like: <div class="oe_chatter" options='{"disable_attachment_box": 1}'>

TaskID: 2144780
Related: https://github.com/odoo/enterprise/pull/8035
Closes: #44568