[REF] web: Control panel Owl
Refactoring of the control panel with the Owl framework. This is where the core
changes have been made. We refactored the MVC structure of the control panel to
a 'component <-> model' relationship. As such:
- Removed the old control panel: view/model/controller/renderer
- Created a new Owl Component (ControlPanel)
- Created a new class: the Model (an Owl Store did not fit here) and a newly
  designed ControlPanelModel extending it

All changes regarding its creation and instantiation can be found in sibling commits.

Co-authored by: Mathieu Duckerts-Antoine <dam@odoo.com>

Part of task 2196029