[IMP] (website_)event(_sale): create leads from attendees
The purpose is to add a new field, main_registration_id, to the event module.
This field allows to save the first registration on other registrations
when a batch of them is created which can occur in website_event, event_sale
and website_event_sale. This is used to link a registration to the good
lead depending on the type of lead creation.

task-2166679