[FIX] website, *: fix theme to custom colors.
* = web_editor

Previously the theme body, text, header, headings and footer colors were
not added to the customs colors of other colorpickers if they were
custom. Now it is the case.
Those colors can be set from a theme color but it will be the css color
of the theme color that will be set not it's name.
This can lead to the custom color being a theme color which was not
checked before as it was not possible. Now we check that!

Also before the theme tab the selected color if it was custom would
always appear in the palette and be marked selected. This is no longer
the case with the like of headings and status colors as their color are
not part of the custom colors available on the color palette. Now we add
the selected color as a custom color to always have the selected marked.

task-2198177