[IMP] base: New groupby field attribute
There are some fields that are quite expensive to group by or to order
by, such field include relational fields (Many2many and One2many) and
binary fields (both large blob of data and filesystem-stored
attachments).

In order to prevent non-efficient operations on those fields, we
introduce a new groupby field attribute. This attribute is a boolean
value controlling whether the field can be grouped by or ordered by.

Task: 2085683