[IMP] crm: add color widget to improve UX
The color picker help the user to choose a color instead of picking it
by id.

The title and arial-label no-color value is remove from the
picker_preview because it's a false if a color is already set, it should be the name of the
current setted color instead.

Task ID 2128166