[IMP] account: define country_code in main view
l10n* modules frequently uses this field to hide
fields not related particular localization country.
So it is better to have this field in main so that
it does not require to define this field in every
localization module where this view is inherited
and country_code field is used in 'attrs' for any
of it's view's element.

Task: 2212790