[IMP] l10n*: make use of country_code field
Now, it is not required to defind a field to get the
country code of current account.move record as it
is defined in main account.move model itself. Also,
it is already defined in parent view, so it is not
required to redefine it even if it is used in attrs.

Task: 2212790