[REV] mrp: unit_factor as computed field
This reverts commit 96b17a82cf1ea1287e0a52c52c78166ef817ce7b.

The issue appears at post inventory. As the reservation on the
stock_move is updated (because the move are split to set some
quantity done), The reserved quantity on quant are so. This is done
via a savepoint. The savepoint will flush the current snapshot. This
flush will recompute all the computed field. This is the unwanted
behavior because all the stock move ready to be validated will have
their unit factor set to 0.

This commit is not exactly the original revert because of the flexible
consumption feature (896352e53b77df2f42de39c0fed897d10ae947b0)

Task : 2212001