[FIX] web: calendar event not order in the right way
Before this commit, events were sorted by ID in the month view.
Events A occurring after event B would be placed on top of B.

We force an event to be "allday" when the view is in month.
So the sequence can't be on the time but on his ID.

After this commit, we don't force the allday flag anymore

Steps to reproduce:

Go in the calendar app
Insert a event at 12:00
Insert a event at 8:00
Insert a event at 14:00
Go in view month the events are not sorted (BUG)

Task ID: 2183249