[FIX] hr_holidays: keep consistent default_get during leave creation
In some case the default_get can return a state confirm
and later for the same record during creation return state draft

You end up with a draft leave altough you saw it 'confirm' when
you were filling the form.

How this is happening ?

During the creation, the web client call default_get
with employee_id in the field_list, this will trigger a specific
order for the search on leave type and thus return state confirm if the first
leave type found has a validation_type != no_validation

When the user save, the employee_id is passed to the create
but not to the default_get that try to fill the missing field
such as state.

Since employee_id is not in fields_list anymore, it's not added
to the context of the search for leave type and you end up with
a different order and thus a different state

Solution:
  Pass the employee_id in the context during creation
  and use it in the default_get if employee_id is not
  computed by the default_get