[FIX] web: end date is not correct on drop (moth view)
due to previous commit, the flag all day is not more present and so some
logic is not applied anymore.

Now we avoid to change the end date of event in FullCalendar

Note this commit change the behaviour on drop event when the view is month.
Before this commit and the previous commit a event of 1H will become an
all day event when is drop to another cell.
After this commit, the time will be the same as the original event only
the date will change