[IMP] web: calendar add a second test for drop in month view
Before this commit, one test was added in commit odoo/odoo@7af13089dffcd3c9d98716af4aa76b923825fdfc
The test is to check in month view that when we move a non all day event
to another day. The event can't change into a all day event.

After this commit, we add a second test because when the all_day mapping
is present the first test fail.