[FIX] web: blank line added before data rows
In following case blank line added:
In tree view with editable="bottom" and sequence handler, with required one
field, conside we have one record and now try to add new record(do not enter
anything to create). Now change the position of the first data row after row
which is in create mode and now just discard the record, this will leads to the
blank row on top of data row.

If the number of rows are more than 4 then no issue at all as we are directly
removing the row but if the rows are less than 4 then we are inserting the
empty row after all the data records to manage the minimum row count of 4,
empty row should always be added at last.

task-2209076

Co-authored-by: Mohammed Shekha <msh@odoo.com>