[FIX] web: don't show all_day slot when no all_day mapping
Before this commit, it was possible to create/move a event is the
all day slot of FullCalendar also when the mapping field wasn't set.

After this commit, now when the all_day mapping is not present we hide
the all day slot in FullCalendar

Adapt also test, due to this change.

Ref:
https://fullcalendar.io/docs/allDaySlot