[FIX] web: avoid to set allDay when the duration of a event is 24H * n
Before this commit, when the duration of a event is 24H * n. THe event is
set as allDay. This is strange to have a event that start a 8H and end
at 8H of next days and the event is in the allDay slot

After this commit, we don't change the allDay flag anymore.

Steps to reproduce:
* Open calendar in week view
* Create a event from 8 -> 16
* Then resize the end of this event from 16 -> 8 of the next day (BUG)