[FIX] web: fix big colleague face
Currently, the Employee picture stays on top of the card of the employee
we're looking for. The image does not disappear after a search is validated.

Purpose of this task is to remove the Employee picture stays on top of the card
of the employee after a search is validated.

In this commit, we disappear the image after a search is validated.

TaskID : 2212808
Closes : #48253