[IMP] sale_timesheet: small improvements for the reporting
Purpose
=======

Small improvements for the reporting in Project Overview and allow to group by date in Project Costs and Revenues report.

Specifications
==============

In Project Overview:
- unfold the details of SO by default
- do not display caret if it is not foldable

In Project Costs and Revenues, the date is retrieved for each account_analytic_line and therefore the values are not summed anymore in the SQL request.