[IMP] utm: Add kanban examples for UTM campaigns
Inspired by what is done in the task kanban of the project module,
add examples of kanban stage setups for UTM campaigns, and allow users
to load them.

There have also been some minor changes to the campaigns view:
- The helper message when there are no UTM campaigns has been changed
to a more descriptive message.
- All stages that existed by default (hardcoded in the data XML) have
been removed, so a user with a fresh database will have no stages in
their kanban. The old hardcoded stages have been moved to the demo data
instead, so they will still be added by default if the user installs
demo data.
- The ghost columns that appear in the kanban when no stages exist have
been renamed to match the old stages that were hardcoded.
- When the user opens the view to create a column, the 'See examples'
link will appear below the 'Column title', just like in the project
addon, and clicking it will open some examples specifically made for UTM
campaigns. Clicking on the 'Use This For My Kanban' button will load the
stages of the selected example.
- The button 'Use This For My Project' in the project examples dialog has the text
'Use This For My Campaigns' in utm campaigns, and will default to 'Use This For My Kanban'
if it is not set in the examples registry.

Task ID: 2047369