[IMP] mail: keep changes of form while schedule the activity
Before this commit, the changes in the form which are done in an edit mode are
erased while the user marks activity as done and schedule a new one in chatter.

After this commit, the changes done in an edit mode will remain the same while
the user marks activity as done and schedule a new one in chatter.

PR https://github.com/odoo/odoo/pull/48434
Task-2160474