[FIX] survey: skip section with no active questions (in conditional questions)
This commit applies conditional questions management on get_next_question
in page_per_section layout.
If going to next (or previous) section, if the section has no question to
display, because none of the questions in this section have been activated
by previous answers, skip the section and look for the next one.

Task ID:
PR#