[IMP] core, base: add tooling to populate database
Some use case like testing performance or upgrade scripts required a database
with prefilled data, covering basic corner cases. A solution can be to
create data a procedural way.

This commit proposes and api to easily populate a model, ususally by giving a
list of possible values for a field or by giving a compute method that will
be based on raw values of other fields.

The basic way to define how to populate a new field is to override _populate_factories,
a method that returned on ordered list of (field, factory).

The definition of a field is a factory, a generate method that will return
an infinite iterator combinning his value(s) with the previous iterator given in parameter.
Some default factories helper are given in tools.populate.py
- iterate(vals, weighs) will ensure that one record is created for each val
by iterating on them, then will random.choice on those vals based on weights
once first iteration is finished.
- cartesian(vals, weights) will make a cartesian product of is own values with previous iterators,
then will start to go random as iterate.
- compute(function) will call the fuction giving the current values and a random object, expecting to
receive a single value for the current field.
- ...

Each iterator returns a dictionnary of values, and the factory should add a
value for the current field(s). The dictionnary also contained a pseudo_field
__complete, that indicate if this iteration is some filling, randomized data
to reach the expected count of record, or a required iteration to create a
corner case. This indicates if a cartesian product is finished, or an iterate
has consumed all values.

The chosen solution of a list of tupple instead of a dict is to ensure that
the order is enforced and to ease insertion.
The order of the factories is quite important, since some compute fields may need other
one to be defined, and cartesian should always be at the begining to avoid to
have to many combination.

Example:
field A: cartesian([T, F])
field B: cartesian([0, 1])
field C: iterate([a, b, c, d, e])
field D: compute(1-B)

_c is shortcut for __complete
_ is a random value, or result of a random value

iter | root  | field A  | field B | field C     | field D   | result
0     {_c:F}  {... A:T}  {... B:0} {...C:a}      {...D:1}    T,0,a,1 complete:False
                         {... B:1} {...C:b}      {...D:0}    T,1,b,0 complete:False
              {... A:F}  {... B:0} {...C:c}      {...D:1}    F,0,c,1 complete:False
                         {... B:1} {...C:d}      {...D:0}    F,1,d,0 complete:False
1     {_c:T}  {... A:_}  {... B:_} {...C:e,_c:F} {...D:_}    _,_,e,_ complete:False
2     {_c:T}  {... A:_}  {... B:_} {...C:_}      {...D:_}    _,_,_,_ complete:True