[IMP] base: warn the user about standard view modification
Purpose of the task is, Warn the user that we do not recommend editing
the architecture of standard views because During migration if the
client does standard views modification then during migration all views
will be updated according to the new odoo version and all customization
will be lost.
So during upgraded database testing issues will come about that views
or client will post issues and ask about where our customization
is gone.

So in this commit, we added alert message for standard view like:
'Editing the architecture of a standard view is not advised, since the
changes will be overwritten during future module updates.
We recommend applying modifications through inherited views or customization
with Odoo Studio.'

TaskID: 2219956
Closes: #48492