[IMP] website, website_blog: perf - use new viewsref
If we need to check multiple views active state to know if the options are
enabled, there is currently no easy way to do that in one go.

This commit introduce `viewsref()` to get most specific views, as does
`viewref()` already does for a single view.

This is useful for `website_blog` where we need to know the state of multiple
options.

At the moment this commit was made, SQL requests went down from 71 to 61.

task-2211013